move the page slot component to the new place in the lib
