+++ projects/storefrontlib/src/cms-structure/index.ts
@@ -1,1 +1,2 @@
+export * from './page/index';
 export * from './seo/index';
