+++ projects/storefrontlib/src/cms-structure/page/index.ts
@@ -1,0 +1,2 @@
+export * from './slot/page-slot.module';
+export * from './slot/page-slot.component';
