+++ projects/storefrontlib/src/cms-structure/page/slot/page-slot.component.spec.ts
@@ -7,8 +7,8 @@ import {
 import { of, Observable } from 'rxjs';
 
 import { PageSlotComponent } from './page-slot.component';
-import { ComponentWrapperDirective } from './component-wrapper.directive';
-import { OutletDirective } from '../../../outlet';
+import { ComponentWrapperDirective } from '../../../lib/cms/components/component-wrapper.directive';
+import { OutletDirective } from '../../../lib/outlet';
 
 class MockCmsService {
   getContentSlot(): Observable<ContentSlotData> {
