+++ projects/storefrontlib/src/cms-structure/page/slot/page-slot.module.ts
@@ -1,0 +1,13 @@
+import { NgModule } from '@angular/core';
+import { CommonModule } from '@angular/common';
+import { PageSlotComponent } from './page-slot.component';
+import { OutletModule } from 'projects/storefrontlib/src/lib/outlet';
+import { CmsModule } from '../../../lib/cms/index';
+
+@NgModule({
+  imports: [CommonModule, OutletModule, CmsModule],
+  providers: [],
+  declarations: [PageSlotComponent],
+  exports: [PageSlotComponent]
+})
+export class PageSlotModule {}
