+++ projects/storefrontlib/src/lib/cms/components/index.ts
@@ -1,7 +1,5 @@
-export * from './page-slot/page-slot.component';
-export * from './page-slot/component-wrapper.directive';
+export * from './component-wrapper.directive';
 
-import { PageSlotComponent } from './page-slot/page-slot.component';
-import { ComponentWrapperDirective } from './page-slot/component-wrapper.directive';
+import { ComponentWrapperDirective } from './component-wrapper.directive';
 
-export const components: any[] = [PageSlotComponent, ComponentWrapperDirective];
+export const components: any[] = [ComponentWrapperDirective];
