+++ projects/storefrontlib/src/lib/cms/page-layout/page-layout.module.ts
@@ -3,9 +3,10 @@ import { CommonModule } from '@angular/common';
 import { PageLayoutComponent } from './page-layout.component';
 import { CmsModule } from '../cms.module';
 import { PageLayoutService } from './page-layout.service';
+import { PageSlotModule } from '../../../cms-structure/page/slot/page-slot.module';
 
 @NgModule({
-  imports: [CommonModule, CmsModule],
+  imports: [CommonModule, CmsModule, PageSlotModule],
   declarations: [PageLayoutComponent],
   providers: [PageLayoutService],
   exports: [PageLayoutComponent]
